- Completing CQRS pattern - Implementing Result pattern - Editing some stuff in the client app

+++ Linkoo.Application/Features/Activity/ActivityCommands/DeleteActivity/DeleteActivityCommandHandler.cs
@@ -3,13 +3,14 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using AutoMapper;
+using Linkoo.Application.Common.Models;
 using Linkoo.Application.Contracts.Persistence.Repositories;
 using Linkoo.Domain.Entities;
 using MediatR;
 
 namespace Linkoo.Application.Features.Activity.ActivityCommands.DeleteActivity
 {
-    public class DeleteActivityCommandHandler : IRequestHandler<DeleteActivityCommand, Unit>
+    public class DeleteActivityCommandHandler : IRequestHandler<DeleteActivityCommand, Result<Unit>>
     {
         private readonly IMapper _mapper;
         private readonly IActivityRepository _activityRepository;
@@ -19,11 +20,17 @@ namespace Linkoo.Application.Features.Activity.ActivityCommands.DeleteActivity
             _mapper = mapper;
             _activityRepository = activityRepository;
         }
-        public async Task<Unit> Handle(DeleteActivityCommand request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(DeleteActivityCommand request, CancellationToken cancellationToken)
         {
-            var activityToDelete = _mapper.Map<Domain.Entities.Activity>(request);
+            var activityToDelete = await _activityRepository.GetByIdAsync(request.Id);
+
+            if(activityToDelete == null) return Result<Unit>.Failure("Failed to delete activity");
+
+            activityToDelete = _mapper.Map<Domain.Entities.Activity>(request);
+
             await _activityRepository.DeleteAsync(activityToDelete);
-            return Unit.Value;
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
